+++ kernel/src/interrupt/apic/mod.rs
@@ -2,3 +2,4 @@
 
 pub mod io;
 pub mod local;
+pub mod pic;
